Split libxprof into libxtimeline and libxtally (#76)
* Add metababel to CI and configure.ac

* add ruby bin to PATH

* add ruby bin to PATH

* install ruby gems globaly

* split libXProf_la_ into libXTimeline_la_ and libXTally_la_

* remove protobuf flags/libs

* clean header

* change plugin name in timeline

* change xprof to tally plugin

* fix babeltrace_thapi

* rename tally c files

* add btx_ prefix to components

* fix timeline
