+++ xprof/btx_tally.h
