+++ xprof/tests/tally.dust.erb
@@ -4,7 +4,7 @@
   :params:
     schemas: [interval_model.yaml]
     trace: interval_<%= namespace[/interval_(.*).dust/,1] %>.yaml
-- :plugin: sink.xprof.tally
+- :plugin: sink.tally.tally
   :name: tally
   :params:
      display: compact
