+++ xprof/timeline.cpp
@@ -1,4 +1,4 @@
-#include "timeline.h"
+#include "btx_timeline.h"
 #include "timeline.hpp"
 #include "xprof_utils.hpp" // typedef
 #include <fstream>
